Extract filters into separate components to reduce duplication
